// apps/erp-alpac-group/src/modules/payroll/ui/pages/collaborator-index/collaborator.page.types.ts
/*   const data = [
        { id: "001", fullName: "Juan Pérez", position: "Desarrollador Senior", department: "Tecnología", status: "Activo", showProfile: "Ver Perfil" },
        { id: "002", fullName: "María García", position: "Analista de RRHH", department: "Recursos Humanos", status: "Activo", showProfile: "Ver Perfil" },
        { id: "003", fullName: "Carlos Rodriguez", position: "Gerente de Ventas", department: "Comercial", status: "Vacaciones", showProfile: "Ver Perfil" },
        { id: "004", fullName: "Lucía Méndez", position: "Diseñadora UX", department: "Producto", status: "Activo", showProfile: "Ver Perfil" },
        { id: "005", fullName: "Roberto Sánchez", position: "Contador", department: "Finanzas", status: "Baja", showProfile: "Ver Perfil" },
    ]
 */

export type Collaborator = {
    id: string
    fullName: string
    position: string
    department: string
    status: string
    showProfile: string
}